Update README.md
Thank you for this awesome little project... I was just about to write something similar and then stumbled across this. Could be something related to the way I installed node but I had to perform the added terminal commands. For anyone else who might have that issue you may wanna add that to the README. Maybe it was just me though!